Cleanup license (MIT)
Most of the code was already "kind of" MIT, somehow the last sentence of
the license was not present in the source files. Also, after the recent
refactoring, most of the new sources didn't have a proper license
header.

This commit adds the proper MIT licensing header and copyright to every
file that didn't have one.

Ccing every contributor according to [1] regardless of their
contribution. If someone opposes to this change, please contact me and
we will solve the problem.

[1] https://github.com/lttng/lttng-analyses/graphs/contributors

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>
Cc: Antoine Busque <antoinebusque@gmail.com>
Cc: Stéphane Graber <stgraber@ubuntu.com>
Cc: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Cc: Yannick Brosseau <yannick.brosseau@gmail.com>
Cc: Philippe Proulx <eeppeliteloop@gmail.com>