Track chronological fd metadata
This commit introduces a chrono_fds object as an attribute of Process
objects. In it, we track when the file (identified by its name)
associated to a given file descriptor has changed. The object
chrono_fds is indexed by file descriptor (number), and each value is
an OrderedDict which stores all the filenames and filetypes associated
with this fd, indexed by timestamp in nanoseconds.

Limitations: if a program opens a file, closes it, and reopens it with
the same file descriptor, without reusing the FD for any other file in
between, the change will not be tracked in chrono_fds.