Refactor IRQ state provider and related classes
The IRQ state provider is now devoid of "analysis" or aggregation code
and now tracks the current state only, and does so using the newly
implemented HardIRQ and SoftIRQ classes in sv. The removed analysis code
will be reimplemented in the IRQ module of the analysis package, once
the notification system between automaton and analysis is implemented.