Refactor and clean-up cli and core
This refactor mainly refactors modules within the cli package in order
to reduce the duplicated code between separate commands. This allows
for new analyses to be added in a much simpler fashion.

The second significant portion of the refactor is the removal of all
refresh/reset or otherwise analysis-state affecting code from the cli,
and its reimplementation within the analysis itself. The cli now
simply feeds events to the analysis (and state automaton), and
receives notifications to which it can react. For now, the only
notification type is received whenever an analysis period has ended,
in which case the default cli behaviour is to output the results.

Some other minor style clean-up is also performed where was deemed
necessary.

Signed-off-by: Antoine Busque <abusque@efficios.com>