Fix: reset analysis when beginning a new period
Previously, the analysis was reset when a period ended. Although the
effect was strictly equivalent when using a single event to mark both
the beginning and end of the period, or a fixed length period, when
using distinct begin and end events the effect is quite different.

Resetting at the end of a period instead of the beginning had the
effect of accumulating analysis results during the periodless gap
before the next begin event and wrongly considering them as part of
the next period. Resetting at the beginning of the next period fixes
this issue.

Signed-off-by: Antoine Busque <abusque@efficios.com>