Add support for different period begin/end keys, constant keys
This allows using differently named fields as a period key for begin
and end events. For instance, comparing the `tid` field from a
`sched_waking` event to that of `next_tid` in `sched_switch`.

This also adds support for constant or fixed key value. That is to
say, with the same TID example, one can specify a particular TID
value, and only periods for which the key matches this value will be
accounted for.

In both new features, the order of fields and values must be
consistent across all arguments for multi-field keys.

Signed-off-by: Antoine Busque <abusque@efficios.com>