Fix: use path to kernel trace to look up tracer version
Currently, only the lttng-modules version is of use for the
analyses. However, multi-domain traces, combining UST and modules,
could lead to unpredictable results wherein the tracer_version was
erroneously identified as that of the userspace tracer, which was
problematic when its version differed from that of the kernel tracer.

Extracting the version directly from the kernel trace's metadata
corrects this issue.

Signed-off-by: Antoine Busque <abusque@efficios.com>