Try to read the text metadata if babeltrace failed
Reading a trace with babeltrace -o ctf-metadata fails if the metadata is
in text format instead of binary format. When we detect this error, try
to read the metadata with 'cat'.

Also, the traces generated with the CTFWriter API can only write strings
in the environment section for now, so we add optional quotes around the
version regex to allow extracting the version in these traces.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>