Generate trace for intersect mode test
This is a basic trace generated to test the intersect mode. The first
three events are in the same stream, the last one is in a different
stream with no overlapping. This trace parsed with babeltrace outputs 4
events by default and 0 with babeltrace --stream-intersection.

Please adapt the test to work with the --disable-intersect-mode.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>