cli: add "catch all" for exceptions to sanitize output
This prevents the user from seeing Python tracebacks, and in MI
mode, it makes sure that only the error object is printed if there is
an error, wherever this error happens.

Eventually, log statements should be added at strategic locations to
optionally provide more information about a failure to the user (and
to bug reports).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>