Sanitize event names when using them as variable/function name
I am trying to use gen_ctfwriter on a trace that contains an event name
"lttng_ust_statedump:start".  It generates code that looks like:

  def write_lttng_ust_statedump:start(self, ...):

Since a colon is not allowed in a function name, it doesn't work.  Fix
this by replacing colons by underscores.  As more problematic special
characters are found, they can be added to the sanitize function.