Add support for nested and parallel periods
Nested and parallel periods are a new feature to isolate the results
of an LTTng analysis within specific periods. Multiple, concurrent
periods are allowed. A period can have a parent period, in which case
its parent must be "active" for it to be active also.

Period definitions are specified on the command line using the
new --period option, of which the argument has the following format:

    [NAME[(PARENT)]] : BEGIN-EXPR [: END-EXPR]

where:

  * NAME: Optional period name.
  * PARENT: Optional parent period's name. The parent must be
    defined using another, preceding --period option.
  * BEGIN-EXPR: Expression that must be matched for a new
    period instance to be created.
  * END-EXPR: Expression that must be matched for the created
    instance to be finished. If this is not specified,
    BEGIN-EXPR is used as END-EXPR also.

An upcoming document will explain the format of the BEGIN-EXPR
and END-EXPR expressions.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>