Test all options with multiple traces
Introduce a new test suite: tests_long_regression that runs all the
scripts we have with most of the options currently defined on the traces
from the lttng_ref_traces repository. The --period arguments are not yet
tested, but will be added soon.

This test only checks the return code of the scripts to detect
unexpected errors, but does not check the actual output produced by the
scripts.

This test is not run by default on setup.py test because it is quite
long (> 30 minutes). To run it manually:
$ ./setup.py test --test-suite test_long_regression

We should add more combination of options and traces, but this is a
start.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>