Split captures in multiple lines
If we capture multiple fields, the output in lttng-periodlog/top can quickly
become hard to read. With this commit, we convert back the value in an
orderedDict and output one capture per line. We don't repeat the timerange and
the rest of the payload so it is clear that all the captured fields belong to
the same period.

Begin                End                   Duration (us) Name  Begin capture      End capture
[10:58:26.157215628, 10:58:26.202139041]       44923.413 wake  $evt.tid = 11159   $evt.next_tid = 11159
                                                               $evt.cpu_id = 1    $evt.cpu_id = 1
                                                                                  $evt.$name = sched_switch

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>