Add aggregation and grouping controls
The new --period-aggregate argument allows to specify a list of periods
to output at the end.

The new --period-aggregate-by controls in which parent period we want to
aggregate the data.

And the new --period-group-by allows to specify a tuple and group all of
the aggregated periods with the same tuple value in the same output.

For now the log and top outputs are working with these new switches.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>