Support for the --group-by option in periodstats/freq
It is now possible to use the --group-by in lttng-periodstats and
lttng-periodfreq to output the tables based on the values of some
captured fields.

For example, after capturing the 'filename' field from the
syscall_entry_open event, it is possible to output the statistics and
frequency distributions based on the value of this field by using the
syntax: --group-by "period.capture", in this case: --group-by
"open.filename".

Multiple captures can separated by a comma. When multiple captures are
specified, the analyses outputs a result for each unique combination of
the captured values.

The MI support is working as well.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>